Rename 'live' to 'dynamic' in code and comments.
