+++ tests/dynamic/test_attribute_template.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
-import prefab_classes.live._attribute_class as attribute_module
-from prefab_classes.live._attribute_class_maker.regenerate_attribute import (
+import prefab_classes.dynamic._attribute_class as attribute_module
+from prefab_classes.dynamic._attribute_class_maker.regenerate_attribute import (
     generate_attribute_source,
 )
